BE-013: Create redirect route when backend method was called by GET
Background:
As user i want to back or refresh in the same page when reload the same page, but in this case. Our page have same return redirect on POST or GET in that page,
so when controller get command to refresh. Laravel page are redirect with POST instead of GET, because route for that backend only exist for POST method. Not GET

Possibility Issue:
- Return issue 405 [Request HTTP are not allowed] when try to switch from mobile to desktop mode or vice versa

Resolution:
- Configure properly backend API that still possibility called on the front-end with GET method to return redirect back to previous page

Signed-off-by: Dicky Herlambang (花） <herlambangdicky5@gmail.com>