Upload: Add new function to count total pdf pages
Since FE are relying to ready this upload / public storage path on BE instances, after
switch to API only route. It seems upload / public storage folder are not accessible from FE
services, due that path are not registered in any route on BE services. Instead directly access to
BE storage folder, create new route and function to count and return exact total PDF pages to FE
services.

Signed-off-by: Dicky Herlambang (花） <herlambangdicky5@gmail.com>