Controllers: Split: Try to handle un-count custom pages
In some condition when custom page ranges was inputted, value sometimes are not
really custom. It can be integer without any string on it, in that case hana-pdf
can threw overflow integer into iLovePDF API.

Although it should be return error from iLovePDF, but it should be good if it can be
handle from backend instead. Apply more guard logic to protect if value was input on
custom page are integer, then do logic to check if page number are not more that total
pdf pages.

And also properly include some database insert on some error, it should be tracked too.

Signed-off-by: Dicky Herlambang (花） <herlambangdicky5@gmail.com>